refactor(application): move ui packages under ui path
