+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/MainScreen.kt
@@ -1,12 +1,12 @@
-package sisterhood.application
+package sisterhood.application.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import sisterhood.application.hentai.HentaiComponent
-import sisterhood.application.hentai.HentaiGridState
+import sisterhood.application.ui.hentai.HentaiComponent
+import sisterhood.application.ui.hentai.HentaiGridState
 import sisterhood.usecase.HentaiInfo
 
 @Composable
