+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/SettingScreen.kt
@@ -1,9 +1,9 @@
-package sisterhood.application.navi
+package sisterhood.application.ui
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun NotFound() {
-    Text("Not Found Route")
+fun SettingScreen() {
+    Text("Hello settings")
 }
