+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/hentai/HentaiComponent.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.hentai
+package sisterhood.application.ui.hentai
 
 import androidx.compose.runtime.Composable
 import sisterhood.usecase.HentaiInfo
