+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/hentai/HentaiGridState.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.hentai
+package sisterhood.application.ui.hentai
 
 import androidx.compose.foundation.lazy.grid.LazyGridState
 import androidx.compose.runtime.*
