+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/hentai/HentaiItem.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.hentai
+package sisterhood.application.ui.hentai
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
