+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/hentai/HentaiThumbnail.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.hentai
+package sisterhood.application.ui.hentai
 
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
