+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/hentai/viewer/HentaiPageList.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.hentai.viewer
+package sisterhood.application.ui.hentai.viewer
 
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
