+++ sisterhood-application/src/commonMain/kotlin/sisterhood/application/ui/navi/NaviScope.kt
@@ -1,4 +1,4 @@
-package sisterhood.application.navi
+package sisterhood.application.ui.navi
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
